Update on "[3/N] Refactor FR script - Add a processor module"
[ghstack-poisoned]
